move card to deployment folder
